requirements.txt
livekit-agents[openai,silero]~=1.0